Refactor tests for unexported functions in tpe
