+++ tpe/export_test.go
@@ -1,0 +1,7 @@
+package tpe
+
+var (
+	ExportGetObservationPairs   = getObservationPairs
+	ExportSplitObservationPairs = (*Sampler).splitObservationPairs
+	ExportSampleCategorical     = (*Sampler).sampleCategorical
+)
